Docker/requirements.txt
stem
PySocks
requests
